requirements.txt
Flask==1.1.1
pytesseract==0.3.0
requests==2.22.0
Pillow==6.2.0
